requirements.txt
python-telegram-bot==13.15
requests
apscheduler==3.10.1
pytz>=2023.3
